changed swagger settings
