+++ src/main/resources/application.properties
@@ -19,7 +19,8 @@ management.endpoint.shutdown.enabled=true
 management.endpoints.web.exposure.include=health,shutdown
 
 #swagger config
-springdoc.swagger-ui.path=/swagger-ui.html
+springdoc.swagger-ui.path=/swagger-ui
 springdoc.api-docs.path=/Prod/v2/api-docs.properties
+
